pyproject.toml
[tool.poetry]
name = "drf-snap-testing"
version = "0.1.0"
description = ""
authors = ["Xavier Francisco <xavier.n.francisco@gmail.com>"]
readme = "README.md"
packages = [
    { include = "drf_snap_testing" },
]


[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
freezegun = "^1.2.2"
vcrpy = "^4.3.1"
drf-yaml = "^3.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
mypy = "^1.3.0"
django-stubs = "^4.2.0"
djangorestframework-stubs = "^3.14.0"
pylint = "^2.17.4"
flake8 = "^6.0.0"
isort = "^5.12.0"
absolufy-imports = "^0.3.1"
black = "^23.3.0"
pytest-random-order = "^1.1.0"


[tool.pytest.ini_options]
python_files = ["tests/*test*.py"]


[tool.black]
target-version = ['py310']


[tool.ruff]
select = [
    "A" , "B", "C", "E", "D", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG",
    "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY",
    "PD", "PGH", "PIE", "PL", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH",
    "TID", "TRY", "UP", "YTT",
]
ignore = [
    "UP006", "UP007", "UP035",
    "ANN101", "ANN401",
    "D203", "D212",
    "TID252",
    "D104", "D100",
]
unfixable = ["ERA001"]

[tool.ruff.per-file-ignores]
"*migrations/*py" = ["ARG001", "N806", "D101", "D103"]


[tool.isort]
profile = "black"


[tool.mypy]
strict = true
disallow_subclassing_any = false

[[tool.mypy.overrides]]
module = ["sqlparse", "vcr"]
ignore_missing_imports = true


[tool.pylint.format]
max-line-length = "88"

[tool.pylint.messages_control]
enable = "useless-suppression"
disable = """
  missing-function-docstring,
  missing-module-docstring,
  missing-class-docstring,
  unsubscriptable-object,
  useless-super-delegation
"""

[tool.pylint.similarities]
min-similarity-lines=10
ignore-imports=true
ignore-comments=false
ignore-docstrings=false
ignore-signatures=false

[tool.pylint.design]
max-parents=12
max-attributes=12
min-public-methods=0


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
